后端/requirements.txt
absl-py==2.1.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
build==1.2.2.post1
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.2.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
Django==5.1.3
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.18.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2025.3.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mysqlclient==2.2.7
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
oss2==2.19.1
package-name==0.1
packaging==24.1
pillow==10.3.0
pip-tools==7.4.1
protobuf==4.25.3
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sqlparse==0.5.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
